Smart Zombies now spawn with new stats

+++ Assets/Scripts/Enemy Scripts/EnemyHealth.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class EnemyHealth : MonoBehaviour {
 
-    public int startingHealth = 100;
-    public int currentHealth;
+    public float startingHealth = 100;
+    public float currentHealth;
     public float sinkSpeed = 2.5f;
 
     private WaveManagement wm;
